Move `wrap` to its own module
