+++ src/core.rs
@@ -1,8 +1,8 @@
 //! Building blocks for advanced wrapping functionality.
 //!
 //! The functions and structs in this module can be used to implement
-//! advanced wrapping functionality when [`wrap()`](super::wrap)
-//! [`fill()`](super::fill()) don't do what you want.
+//! advanced wrapping functionality when [`wrap()`](crate::wrap())
+//! [`fill()`](crate::fill()) don't do what you want.
 //!
 //! In general, you want to follow these steps when wrapping
 //! something:
