+++ src/fuzzing.rs
@@ -10,7 +10,7 @@ pub fn fill_slow_path<'a>(text: &str, options: Options<'_>) -> String {
 
 /// Exposed for fuzzing so we can check the slow path is correct.
 pub fn wrap_single_line<'a>(line: &'a str, options: &Options<'_>, lines: &mut Vec<Cow<'a, str>>) {
-    super::wrap_single_line(line, options, lines);
+    crate::wrap::wrap_single_line(line, options, lines);
 }
 
 /// Exposed for fuzzing so we can check the slow path is correct.
@@ -19,5 +19,5 @@ pub fn wrap_single_line_slow_path<'a>(
     options: &Options<'_>,
     lines: &mut Vec<Cow<'a, str>>,
 ) {
-    super::wrap_single_line_slow_path(line, options, lines)
+    crate::wrap::wrap_single_line_slow_path(line, options, lines)
 }
